feat(quiz) location modified, attemptid removed, useclient removed, component imported. (#FE-14)
